chore(kit): move all common date-utilities to `utils/date`-folder
