+++ projects/kit/src/lib/utils/date/date-to-segments.ts
@@ -1,4 +1,4 @@
-import {MaskitoDateSegments} from '../types';
+import {MaskitoDateSegments} from '../../types';
 
 export function dateToSegments(date: Date): MaskitoDateSegments {
     return {
