+++ projects/kit/src/lib/utils/index.ts
@@ -1,11 +1,11 @@
 export * from './clamp';
-export * from './date-segment-value-length';
-export * from './date-to-segments';
+export * from './date/date-segment-value-length';
+export * from './date/date-to-segments';
+export * from './date/is-date-string-completed';
+export * from './date/parse-date-string';
+export * from './date/segments-to-date';
+export * from './date/to-date-string';
+export * from './date/validate-date-string';
 export * from './get-object-from-entries';
-export * from './is-date-string-completed';
 export * from './pad-with-zeroes-until-valid';
-export * from './parse-date-string';
-export * from './segments-to-date';
 export * from './split-into-chunks';
-export * from './to-date-string';
-export * from './validate-date-string';
